Secret reconciler
When Claudie creates a Kubernetes cluster, a secret is created containing
the kubeconfig to access the cluster.
Secret reconciler watches for Claudie's secrets and:
1. creates a SveltosCluster when a secret is created;
2. deletes corresponding SveltosCluster when secret is deleted.

This simplifies Claudie/Sveltos integration.
By creating a SveltosCluster, any Claudie created cluster is now knwon
to Sveltos which can then automatically manage add-ons and applications.